T195077/docker compose for tests (#253)
* Add ORES docker-compose file for testing purposes

Services:
* ores-api
* ores-worker
* redis

Considerations:
* For simplicity, both the task queue and the score cache are pointed at
the same redis. It can be split into 2 when needed.
* I was skeptical about adding changes to test config files without
breaking anything, so I added a new set of config files
'celery-test-docker'. It adds a 'docker_celery' to scoring_systems, a
'docker_redis' to score_caches and sets ores to use 'docker_celery' as
its scoring system.
* Celery seems to complain very hard when the service is run as root
(which makes sense). I set the env var C_FORCE_ROOT to true to get it to
start. Probably the best would be adding the user to application is
supposed to be run by to the docker image.

* Adapt Makefile to docker-compose