swagger: update references to what event type is needed for precache
It seems that the v3 swagger UI suggests to use a recentchange event
as input for the precache endpoint, but after trying it I got only
errors related to the field "database" not present in the event's data.

After trying with a revision create event, I was able to get
the scores as expected.

The ChangeProp code responsible to send the event to ORES
seems to agree as well:
https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/services/change-propagation/+/refs/heads/master/sys/ores_updates.js#121

Bug: T301878